dev: updated `std` packages added in 1.21

+++ scripts/generate-std.sh
@@ -2,11 +2,15 @@
 
 mkdir -p .tmp
 
+tip=21
+
 # create std pkg list
-for i in $(seq 1 20); do
-    docker run --rm -it golang:1.$i go list std \
-        | grep -v internal | grep -v vendor > .tmp/go@1.$i;
-    docker rmi golang:1.$i
+for i in $(seq 1 $tip); do
+    if [[ ! -a .tmp/go@1.$i ]]; then
+        docker run --rm -it golang:1.$i go list std \
+            | grep -v internal | grep -v vendor > .tmp/go@1.$i;
+        docker rmi golang:1.$i
+    fi
 done
 
 dest="analyzer/std.go"
@@ -21,7 +25,7 @@ echo "package analyzer" >> $dest
 echo "" >> $dest
 echo "var std = map[string]struct{}{" >> $dest
 
-for i in $(seq 2 20); do
+for i in $(seq 2 $tip); do
 
     printf "\t// added in Go v1.%s in compare to v1.%s (docker image)\n" $i "$(($i-1))" >> $dest
